# Refactor metrics to move unit conversion logic to otel_metrics