Metrics units should be a concern of metrics related files as opposed to the code that generates those metrics.

### Description

Perf:

| Branch | File Size  |   Read BW    |  Write BW  | RandRead BW  | RandWrite BW |
|--------|----------|--------------|-----------|----------------|----------------|
| Master |  0.25MiB   | 553.65MiB/s  | 1.09MiB/s  |  81.91MiB/s  |  1.14MiB/s   |
|   PR   |  0.25MiB   | 573.52MiB/s  | 1.18MiB/s  |  81.43MiB/s  |  1.22MiB/s   |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 48.828MiB  | 3923.0MiB/s  | 84.98MiB/s | 1518.44MiB/s |  85.93MiB/s  |
|   PR   | 48.828MiB  | 3930.72MiB/s | 82.77MiB/s | 1557.12MiB/s |  84.43MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 976.562MiB | 3875.87MiB/s | 79.38MiB/s | 687.03MiB/s  |  30.76MiB/s  |
|   PR   | 976.562MiB | 3874.53MiB/s | 79.33MiB/s | 942.37MiB/s  |  29.43MiB/s  |

### Link to the issue in case of a bug fix.
b/417142000

### Testing details
1. Manual - Manually verified that the latency metrics are of the same order before and after the change.
2. Unit tests - NA
3. Integration tests - Tested by existing integration tests in prom_test.go

### Any backward incompatible change? If so, please explain.
No